+++ changelog.md
@@ -4,6 +4,7 @@
 <!-- Add all new changes here. They will be moved under a version at release -->
 * `New` Omit parameter hints when the argument name matches
 * `FIX` Fix a typo in `no-unknown` diagnostic message
+* `FIX` Autodoc generation so it does not include documentation for builtin Lua language features
 
 ## 3.16.1
 `2025-12-8`